+++ package.json
@@ -1,7 +1,6 @@
 {
-  "name": "@test/source",
+  "name": "envio-vital-front",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -25,60 +24,31 @@
     "@angular/router": "^18.2.0",
     "@angular/ssr": "^18.2.1",
     "@ngx-translate/core": "^15.0.0",
-    "@nx/angular": "19.4.1",
-    "@splitsoftware/splitio": "^10.27.0",
-    "geolib": "^3.3.4",
-    "moment": "^2.30.1",
-    "ol": "^9.2.4",
+    "autoprefixer": "^10.4.20",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~18.0.0",
+    "@angular-devkit/build-angular": "^18.2.1",
     "@angular-devkit/core": "~18.0.0",
     "@angular-devkit/schematics": "~18.0.0",
-    "@angular-eslint/eslint-plugin": "^18.0.1",
-    "@angular-eslint/eslint-plugin-template": "^18.0.1",
-    "@angular-eslint/template-parser": "^18.0.1",
-    "@angular/cli": "~18.0.0",
-    "@angular/compiler-cli": "~18.0.0",
+    "@angular/cli": "^18.2.1",
+    "@angular/compiler-cli": "^18.2.0",
     "@angular/language-service": "~18.0.0",
-    "@angular/material": "^18.1.0",
-    "@nx/devkit": "19.4.1",
-    "@nx/eslint": "19.4.1",
-    "@nx/eslint-plugin": "19.4.1",
-    "@nx/jest": "19.4.1",
-    "@nx/js": "19.4.1",
-    "@nx/playwright": "19.4.1",
-    "@nx/storybook": "^19.4.1",
-    "@nx/web": "19.4.1",
-    "@nx/workspace": "19.4.1",
-    "@playwright/test": "^1.36.0",
     "@schematics/angular": "~18.0.0",
-    "@storybook/addon-essentials": "7.6.20",
-    "@storybook/addon-interactions": "^7.5.3",
-    "@storybook/angular": "7.6.20",
-    "@storybook/core-server": "7.6.20",
-    "@storybook/jest": "^0.2.3",
-    "@storybook/test-runner": "^0.13.0",
-    "@storybook/testing-library": "^0.2.2",
-    "@swc-node/register": "~1.9.1",
-    "@swc/core": "~1.5.7",
-    "@swc/helpers": "~0.5.11",
-    "@types/jest": "^29.4.0",
-    "@types/node": "^18.16.9",
-    "@typescript-eslint/eslint-plugin": "^7.3.0",
-    "@typescript-eslint/parser": "^7.3.0",
-    "@typescript-eslint/utils": "^8.0.0-alpha.28",
-    "eslint": "~8.57.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-playwright": "^0.15.3",
-    "ngx-electron": "^2.2.0",
-    "nx": "19.4.1",
-    "storybook": "^7.5.3",
-    "ts-jest": "^29.1.0",
-    "ts-node": "10.9.1",
-    "typescript": "~5.4.2"
+    "@types/express": "^4.17.17",
+    "@types/jasmine": "~5.1.0",
+    "@types/node": "^18.18.0",
+    "autoprefixer": "^10.4.20",
+    "jasmine-core": "~5.2.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.4.14",
+    "typescript": "~5.5.2"
   }
 }
