+++ src/app/components/almacen-view/almacen-view.component.html
@@ -56,5 +56,5 @@
       </div>
     </div>
   </div>
-      <app-lista-eventos [eventos]="eventos"></app-lista-eventos>
+      <app-lista-eventos-almacen [eventos]="eventos"></app-lista-eventos-almacen>
 </div>
