+++ src/app/components/lista-almacenes/lista-almacenes.component.ts
