+++ src/app/components/tarjeta-almacen/tarjeta-almacen.component.ts
