+++ src/app/components/tarjeta-evento/tarjeta-evento.component.spec.ts
