+++ src/app/interfaces/evento.ts
@@ -1,0 +1,8 @@
+export interface Evento {
+    id: number;
+    idEvento: number;
+    nombreEvento: string;
+    descripcionEvento: string;
+    nombreProvincia: string;
+    estado: boolean;
+  }
